refactor(list-devices): Create a `list_devices` mod
The corresponding code for `bt list-devices` is moved under its own
module to keep lib root lean and achieve modularity within `bt`.
